requirements.txt
streamlit
face_recognition
opencv-python
pandas
Pillow